+++ packages/melos/lib/src/commands/runner.dart
@@ -8,7 +8,6 @@ import 'package:cli_util/cli_logging.dart';
 import 'package:collection/collection.dart';
 import 'package:conventional_commit/conventional_commit.dart';
 import 'package:file/local.dart';
-import 'package:graphs/graphs.dart';
 import 'package:mustache_template/mustache.dart';
 import 'package:path/path.dart';
 import 'package:pool/pool.dart';
@@ -23,6 +22,7 @@ import '../common/git_commit.dart';
 import '../common/glob.dart';
 import '../common/pending_package_update.dart';
 import '../common/platform.dart';
+import '../common/pub_dependency_list.dart';
 import '../common/utils.dart' as utils;
 import '../common/utils.dart';
 import '../common/versioning.dart' as versioning;
