+++ packages/melos/test/commands/publish_test.dart
@@ -1,6 +1,6 @@
 import 'package:graphs/graphs.dart';
 import 'package:melos/melos.dart';
-import 'package:melos/src/commands/runner.dart';
+import 'package:melos/src/common/utils.dart';
 import 'package:pub_semver/pub_semver.dart';
 import 'package:pubspec/pubspec.dart';
 import 'package:test/expect.dart';
